Membuat migration school dan menambahkan endpoint get dan post school

+++ controllers/ebookController.js
@@ -51,6 +51,7 @@ const createEbook = async (req, res) => {
     const {categoryId, name, description, writer, publisher, publicationYear, isbn} = req.body;
     const decoded = jwt_decode(req.cookies.accessToken);
     const errors = validationResult(req);
+    const user = await User.findOne({where: {id: decoded.userId}});
 
     if (!errors.isEmpty()) {
         if (req.files) {
@@ -60,20 +61,38 @@ const createEbook = async (req, res) => {
         return res.status(400).json({errors: errors.array()});
     }
 
-    await Ebook.create({
-        userId: decoded.userId,
-        categoryId,
-        name,
-        image: req.files.ebookImage[0].filename,
-        fileName: req.files.ebookFile[0].filename,
-        description,
-        writer,
-        publisher,
-        publicationYear,
-        isbn,
-    });
+    if (!user || user.statusId == 0 || user.roleId == 3) {
+        if (req.files) {
+            fs.unlinkSync(req.files.ebookImage[0].path);
+            fs.unlinkSync(req.files.ebookFile[0].path);
+        }
+        return res.status(401).json({msg: "Unauthorized"});
+    }
 
-    return res.status(200).json({msg: "Ebook berhasil diupload"});
+    try {
+        if (req.files.ebookImage) {
+            await Ebook.create({
+                image: req.files.ebookImage[0].filename,
+            });
+        }
+
+        await Ebook.create({
+            userId: decoded.userId,
+            categoryId,
+            name,
+            fileName: req.files.ebookFile[0].filename,
+            description,
+            writer,
+            publisher,
+            publicationYear,
+            isbn,
+        });
+
+        return res.status(200).json({msg: "Ebook berhasil diupload"});
+    } catch (err) {
+        console.log(err);
+        return res.sendStatus(500);
+    }
 };
 
 const editEbook = async (req, res) => {
@@ -148,10 +167,16 @@ const updateEbook = async (req, res) => {
 };
 
 const deleteEbook = async (req, res) => {
-    const {id} = req.body;
+    const id = req.params.id;
+    const decoded = jwt_decode(req.cookies.accessToken);
 
     try {
-        const ebook = await Ebook.findOne({where: {id}, attributes: ["image", "fileName"]});
+        const ebook = await Ebook.findOne({where: {id}, attributes: ["userId", "image", "fileName"]});
+
+        if (ebook.userId != decoded.userId) {
+            return res.status(403).json({msg: "Forbiden"});
+        }
+
         await Ebook.destroy({
             where: {
                 id,
